feat(chat-webview): extract pure UI package from chat-sidebar
Greenfield extraction, second step in the three-tier split. The
webview was already 100% vscode-free; this commit relocates it to a
dedicated published-shape package and introduces a clean
transport-agnostic bridge contract.

What moved (via git mv, preserving history):
- packages/chat-sidebar/src/webview/* → packages/chat-webview/src/

What changed in the UI:
- Protocol types (CompletedTurn, ContextSummary, HostToWebview,
  WebviewToHost, ProactiveSuggestion) now come from @lace/chat-core,
  re-exported from chat-webview for consumer convenience.
- App's `vscode: VsCodeApi` prop renamed to `bridge: WebviewBridge` —
  JetBrains (or any other host) provides the same contract
  unchanged. JCEF's `postMessage` shim + `window.message` events
  match exactly.

What chat-sidebar now owns:
- `src/webview-entry.tsx` — VS Code-specific bootstrap that calls
  `acquireVsCodeApi()`, wraps it in WebviewBridge, and mounts
  `<App bridge={...} />` from @lace/chat-webview. Rspack entry
  updated.

Nothing on the host side changed yet. chat-sidebar/src/host/ still
runs the old code path unchanged. Its local protocol.ts is now a
duplicate of chat-core/src/protocol.ts — to be collapsed in the
adapter-rewrite commit.

Verified: tsc green, unit tests green (123 pass), rspack builds all
three bundles successfully, biome clean.
