--- packages/chat-sidebar/src/webview/global.d.ts
@@ -1,9 +1,0 @@
-declare global {
-  function acquireVsCodeApi(): {
-    postMessage(msg: unknown): void;
-    getState(): unknown;
-    setState(state: unknown): void;
-  };
-}
-
-export {};
