--- packages/chat-sidebar/src/webview/index.tsx
@@ -1,9 +1,0 @@
-import { createRoot } from 'react-dom/client';
-import { App } from './App';
-import './styles.css';
-
-const vscode = acquireVsCodeApi();
-const container = document.getElementById('root');
-if (container) {
-  createRoot(container).render(<App vscode={vscode} />);
-}
